Rudimentory error handling
Responds with a generic error message if something goes wrong. I don't think I
want to go down the rabbit hole and try to analyze the error messages at this
point. Given that we need a temporary path to begin with the rendering, the
Either idea doesn't really work unless I want to do all the bracketing myself.

Fix #3